requirements.txt
click==8.1.3
Flask==2.2.2
hedera-sdk-py==2.17.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
JPype1==1.4.0
MarkupSafe==2.1.1
pyjnius==1.4.2
python-dotenv==0.21.0
six==1.16.0
Werkzeug==2.2.2
